cfg: Move ConfigFile classes to ini.cpp/h
