+++ core/cfg/cfg.h
@@ -19,42 +19,3 @@ s32  cfgExists(const wchar * Section, const wchar * Key);
 void cfgSetVirtual(const wchar * lpSection, const wchar * lpKey, const wchar * lpString);
 
 bool ParseCommandLine(int argc,wchar* argv[]);
-
-struct ConfigEntry
-{
-	u32 flags;
-	string name;
-	string value;
-	string valueVirtual;
-	ConfigEntry* next;
-	ConfigEntry(ConfigEntry*);
-	string GetValue();
-	void SaveFile(FILE*);
-};
-
-struct ConfigSection
-{
-	u32 flags;
-	string name;
-	ConfigEntry* entrys;
-	ConfigSection* next;
-	~ConfigSection();
-	ConfigSection(ConfigSection*);
-	ConfigEntry* FindEntry(string);
-	void SetEntry(string, string, u32);
-	void SaveFile(FILE*);
-};
-
-struct ConfigFile
-{
-	ConfigSection* entrys;
-	~ConfigFile();
-	void ParseFile(FILE*);
-	void SaveFile(FILE*);
-	ConfigSection* FindSection(string);
-	ConfigSection* GetEntry(string);
-	s32 Exists(const wchar *, const wchar *);
-	void  LoadStr(const wchar *, const wchar *, wchar *,const wchar*);
-	string  LoadStr(const wchar *, const wchar *, const wchar*);
-	s32  LoadInt(const wchar *, const wchar *,s32);
-};
